requirements.txt
numpy
tensorflow
matplotlib
scikit-learn
Pillow
opencv-python
seaborn